hgBlacklist {seqDNAcopy}	R Documentation

Blacklisted regions for human genome

Description

Regions in the genome that give sequencing artifacts as listed in UCSC Genome browser track wgEncodeDacMapabilityConsensusExcludable

Usage

  data(hgBlacklist)

Format

A list containing 25 matrices corresponding to chromosomes 1-22, X, Y & MT in that order. Each row is an interval that is to be discarded.

Details

The intervals for hg19 were obtained from UCSC Genome browser track https://genome.ucsc.edu/cgi-bin/hgFileUi?db=hg19&g=wgEncodeMapability wgEncodeDacMapabilityConsensusExcludable. The intervals were padded by 100 bases on both ends and an extra row (2^28, 2^28) added to each chromosome for programming convenience. The hg18 and hg38 data were obtained using liftover https://genome.ucsc.edu/cgi-bin/hgLiftOver.

A fragment whose mid-point lies in an interval is discarded from analysis.

References

https://sites.google.com/site/anshulkundaje/projects/blacklists

[Package seqDNAcopy version 0.0.3 Index]