seqDNAcopy-package {seqDNAcopy}	R Documentation

Copy number data analysis using DNA sequencing data

Description

Implements the CBS algorithm for sequencing data. (Update for Mouse [nds])

Details

seqDNAcopy implements the CBS algorithm for sequencing data. It uses coverage depth for tumor and normal data to compute log-ratio which is first normalized for GC content and then segmented using segment function from DNAcopy.

Author(s)

Venkatraman E. Seshan

Maintainer: Venkatraman E. Seshan <seshanv@mskcc.org>

References

Need to be added

[Package seqDNAcopy version 0.0.3 Index]